Revert "changes"
This reverts commit cc904b1f12ddb3216091e1b555a9d42fa9630f5e.

+++ Utils/Utilities.c
@@ -14,6 +14,8 @@
 #include "esp_err.h"
 #include "esp_log.h"
 
+#include "Utilities.h"
+
 /****** Function Prototypes ***********/
 
 /************ ISR *********************/
@@ -25,7 +27,7 @@
 /****** Global Data *******************/
 
 /****** Global Functions *************/
-void showmem(uint8_t *memptr, int len)
+void showmem(uint8_t *memptr, uint16_t len)
 {
     for (int i = 0; i < len; i++)
     {
